fix: Rebuild AIF prediction network.
The previous network did not work for unknown reasons. It could be,
obscurely, that the final output layer and the labels were transposed
from each other, but the cost could be calculated because of broadcasting
anyway. It could also be that I was using too few nodes, or too high a
learning rate. A low learning rate and some time to get out of the average
 value local minima were necessary to get this network fitting. It is pretty
 messy currently, and the old sheri_network will likely no longer work, but
it can be built from.

Coming up from here: dynamic_rnn implementation, and tests to see how much
noise one can add to the network. It would be nice to have a penalty for
low-variance networks -- google that?